requirements.txt
Flask==3.1.0
inference==0.29.0
plotly==5.24.1
PyYAML==6.0.2
